Rename UserEditPage -> EditUserPage and move into Account module for consistency
